% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-plot-add_text_to_plot.R
\name{add_text_to_plot}
\alias{add_matrix_text}
\alias{add_text_to_plot}
\alias{add_vec_text}
\title{Add summary statistics to a plot}
\usage{
add_matrix_text(Matrix.plot, Matrix_text, p, verbose = FALSE, num_level = 0)

add_vec_text(vec.plot, vec_text, verbose = FALSE, num_level = 0)
}
\arguments{
\item{Matrix.plot}{A \code{ggplot} plot faceted with \code{facet_grid(i ~k)}}

\item{Matrix_text}{A list of summary statistics for each parameter index. This
should be in the same format as that returned by
\code{\link{get_Matrix_text}}}

\item{p}{The dimension of the matrix}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}

\item{vec.plot}{A \code{ggplot} plot faceted with \code{facet_wrap(~ i)}}

\item{vec_text}{A list of summary statistics for each
parameter index. This should be in the same format as that returned by
\code{\link{get_vec_text}}.}
}
\value{
Returns a \code{ggplot} plot with the text added
}
\description{
Add summary statistics to a plot
}

