% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-run-adjust_zeros.R
\name{adjust_zeros}
\alias{adjust_zeros}
\title{Replace all zeros in a data frame with a small constant; possibly
  renormalize by samples}
\usage{
adjust_zeros(data, renorm = TRUE, zero_adj = 1e-04, verbose = FALSE,
  num_level = 0)
}
\arguments{
\item{data}{The data frame to be adjusted; assumed to have rows for samples}

\item{renorm}{Whether to renormalize the data after adjusting the zeros}

\item{zero_adj}{The amount to use instead of zer.}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Replace all zeros in a data frame with a small constant; possibly
  renormalize by samples
}

