% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banocc.R
\docType{package}
\name{banocc}
\alias{banocc}
\alias{banocc-package}
\title{banocc: A package for Bayesian ANalysis of Compositional Correlation}
\description{
This package contains many helpful functions for analyzing data this way, as
  well as other useful functions for my personal use.
}
\section{banocc functions}{

ADD FUNCTIONS HERE
}

