% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banocc_data}
\alias{banocc_data}
\title{Simulated relative abundances for 9 features across
100 samples.}
\format{A data frame with 100 rows (the samples)
  and 9 variables (the features)}
\usage{
banocc_data
}
\description{
Simulated relative abundances for 9 features across
100 samples.
}
\keyword{datasets}

