% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes-model-run_model.R
\docType{data}
\name{bayesStanModel}
\alias{bayesStanModel}
\title{The stan model used in the Bayesian fit}
\format{\preformatted{ chr "data \{\\n    int<lower=0> P; // number of elements of the composition\\n    int<lower=0> N; // the number of subjects sampled\\n  "| __truncated__
}}
\usage{
bayesStanModel
}
\description{
This is the literal model used for fitting in Stan
}
\keyword{datasets}

