% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_min_width.R
\name{check_CI}
\alias{check_CI}
\title{Checks a credible interval list (each elt is an array) against
  a null value; returns a list. Each value in a list elt is
  TRUE if the CI includes null_value, FALSE if it doesn't and NA
  if the CI width is 0}
\usage{
check_CI(CI, null_value, verbose = FALSE, num_level = 0)
}
\description{
Checks a credible interval list (each elt is an array) against
  a null value; returns a list. Each value in a list elt is
  TRUE if the CI includes null_value, FALSE if it doesn't and NA
  if the CI width is 0
}

