% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_stan_calc.R
\name{check_stan_lik}
\alias{check_stan_lik}
\title{Check that the log-posterior as calculated by stan is approximately correct}
\usage{
check_stan_lik(res, iter)
}
\arguments{
\item{res}{The printed output from \code{\link[rstan]{sampling}} as a
character vector; each element is a single line of output.}

\item{iter}{The iteration of stan from which to compare (since multiple
  iterations can be present in \code{res})}
}
\description{
Prints out the change in log-posterior after each parameter subtracted from
  the value as calculated by an R function
}

