% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_corrmat_figure.R
\name{compare_two_corrmat_figure}
\alias{compare_two_corrmat_figure}
\title{Generate a heatmap figure comparing two correlation matrices}
\usage{
compare_two_corrmat_figure(Rho.lower, Rho.upper, title = NA,
  truestatus_mat = NA, verbose = FALSE, categories = c("Truth",
  "Estimate"), corr_label = expression(rho["jj'"]), include_text = TRUE,
  text.size = 20, corr.text.size = 6, legend.position = "horizontal",
  num_level = 0)
}
\arguments{
\item{Rho.lower}{The lower correlation matrix}

\item{Rho.upper}{The upper correlation matrix; must match \code{Rho.lower} in dimension}

\item{title}{The title of the plot}

\item{truestatus_mat}{An optional matrix indicating by Boolean values which
elements of Rho.lower and Rho.upper should be non-zero.  The dimensions
must match those of Rho.lower and Rho.upper.}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{categories}{A character vector of length two. The first element
indicates the label for the lower triangle; the second indicates the label
for the upper triangle.}

\item{corr_label}{The label for the legend (can be an expression or a
string).}

\item{include_text}{Boolean: should the values of the correlation matrix be
added as text to the plot?}

\item{text.size}{The base size of the text for \code{ggplot} (in pt size).}

\item{legend.position}{One of either \code{"horizontal"} or
\code{"vertical"}; indicates whether the legends will be positioned at the
bottom (\code{"horizontal"}) or left (\code{"vertical"}) of the plot.}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Generate a heatmap figure comparing two correlation matrices
}
\examples{
Rho.lower <- matrix(c(1, 0, 0, 0.5, 0,
                      0, 1, 0, 0, 0,
                      0, 0, 1, 0, 0,
                      0.5, 0, 0, 1, -0.5,
                      0, 0, 0, -0.5, 1), ncol=5)
X <- rmvnorm(5, rep(0, 5), Rho.lower)
Rho.upper <- cov2cor(X \%*\% t(X))
truestatus_mat <- as.logical(Rho.lower != 0 - diag(5))
title <- "Estimating based on 5 samples"
categories <- c("Truth", "Estimate (5 samples)")
compare_two_corrmat_figure(Rho.lower, Rho.upper, title, truestatus_mat,
                           categories=categories)
compare_two_corrmat_figure(Rho.lower, Rho.upper, title, truestatus_mat,
                           categories=categories, legend.position="vertical")

}

