% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_two_varmat_table.R
\name{compare_two_varmat_table}
\alias{compare_two_varmat_table}
\title{Make a latex table comparing one symmetric matrix with another}
\usage{
compare_two_varmat_table(Mat.lower, Mat.upper)
}
\arguments{
\item{Mat.lower}{One symmetric matrix}

\item{Mat.upper}{Second symmetric matrix}
}
\value{
Returns a string that, when used with \code{cat}, forms a LaTeX table
  comparing \code{Mat.lower} (in the lower triangle) with
  \code{Mat.upper} (in the upper triangle).
}
\description{
Make a latex table comparing one symmetric matrix with another
}

