% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{convert_params}
\alias{convert_params}
\alias{get_ln_params}
\alias{get_n_params}
\title{Convert normal and lognormal mean/variance parameters}
\usage{
get_n_params(mean, Var)

get_ln_params(mean, Var)
}
\arguments{
\item{mean}{the numeric mean of the lognormal or
normal distribution; for the lognormal distribution, must be
greater than zero.}

\item{Var}{the numeric variance-covariance matrix of the lognormal or normal
distribution.}
}
\value{
A named list: \code{_mean} is the converted mean and \code{_Var}
  is the converted variance-covariance matrix.
}
\description{
Convert normal and lognormal mean/variance parameters
}
\examples{
get_n_params(rep(1, 4), diag(4))
get_ln_params(rep(1, 4), diag(4))
get_n_params(get_ln_params(rep(0, 4), diag(4)))

}

