% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_hard_null}
\alias{counts_hard_null}
\title{Simulated count data with no feature correlations}
\format{A data frame with 1000 rows (samples) and 9 variables (the
  features)}
\usage{
counts_hard_null
}
\value{
A data frame with 1000 unconstrained samples from 9 features.
}
\description{
Nine features are draw independently from very different log-normal
  distributions whose means and variances are positively correlated. This
  means that the compositions generated from this dataset
  (see \code{compositions_hard_null}) should be have a correlation
  structure very different from that of these counts.
}
\keyword{datasets}

