% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_neg_spike}
\alias{counts_neg_spike}
\title{Simulated count data with one negative feature correlation}
\format{A data frame with 1000 rows (samples) and 9 variables (the
  features)}
\usage{
counts_neg_spike
}
\value{
A data frame with 1000 unconstrained samples from 9 features.
}
\description{
Nine features are drawn from a log-normal distribution with one negative
  correlation. The resulting compositions are in
  \code{compositions_neg_spike}
}
\keyword{datasets}

