% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_null}
\alias{counts_null}
\title{Simulated count data with no feature correlations}
\format{A data frame with 1000 rows (the samples)
  and 9 variables (the features)}
\usage{
counts_null
}
\value{
A data frame with 1000 unconstrained samples from 9 features.
}
\description{
Nine features are drawn independently from similar log-normal
  distributions to generate null count data. Because the feature
  distributions are very similar, the compositions generated from
  this dataset (see \code{compositions_null}), should have a correlation
  structure similar to that of the counts.
}
\keyword{datasets}

