% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_min_width.R
\name{eval_width}
\alias{eval_width}
\title{Evaluate everything for a particular width and update min_width and
  missing_min_width lists accordingly.}
\usage{
eval_width(min_width, missing_min_width, posterior_samples, parameter.names,
  width, type, null_value, verbose = FALSE, num_level = 0)
}
\arguments{
\item{min_width}{A list; each elt is a parameter. The values in an element
are the minimum widths.}

\item{missing_min_width}{The same structure as min_width, but the values
in an element are indicators of whether the min_width has been achieved.}

\item{posterior_samples}{A list of the posterior samples for each parameter.
Each element should be an array of the samples across all chains for a
parameter. The first dimension of the array should be the sample id; the
others, the parameter index.}

\item{parameter.names}{A list of parameters for which estimates or
quantiles will be returned}

\item{width}{The current width of credible interval to evaluate}

\item{type}{One of \code{"marginal.centered"} or \code{"marginal.hpd"},
indicating whether centeral or highest posterior density intervals are
desired}

\item{null_value}{The null value of the ``hypothesis test''}

\item{verbose}{Print informative statements as the function executes?xo}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Evaluate everything for a particular width and update min_width and
  missing_min_width lists accordingly.
}

