% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-output-get_min_width.R
\name{find_idx_to_update}
\alias{find_idx_to_update}
\title{Find which elements need to have the min_width updated. These are those
  whose CIs include the null (CI_include_null) AND haven't yet
  been updated (missing_min_width)}
\usage{
find_idx_to_update(CI_include_null, missing_min_width, verbose = FALSE,
  num_level = 0)
}
\description{
Find which elements need to have the min_width updated. These are those
  whose CIs include the null (CI_include_null) AND haven't yet
  been updated (missing_min_width)
}

