% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-figures-generate_estimate_figure.R
\name{generate_estimate_figure}
\alias{generate_estimate_figure}
\alias{generate_estimate_figure_MCMC}
\alias{generate_estimate_figure_optim}
\title{Generate a heatmap figure comparing the posterior fits from a \code{stan}
  fit or optimization object.}
\usage{
generate_estimate_figure_MCMC(Fit, A, estimate_method = "median", thin = 1,
  verbose = FALSE)

generate_estimate_figure_optim(Fit, A, Data)
}
\arguments{
\item{Fit}{Either a \code{stanfit} or a list with element \code{par} being
a named vector of parameter estimates}

\item{A}{The dataset of true counts (absolute abundances)}

\item{estimate_method}{Either \code{"mean"} or \code{"median"}, indicating
which estimate to use.}

\item{verbose}{Boolean: print informative statements as
function runs?}
}
\description{
Generate a heatmap figure comparing the posterior fits from a \code{stan}
  fit or optimization object.
}

