% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-figures-generate_inference_figure.R
\name{generate_inference_figure}
\alias{generate_inference_figure}
\alias{generate_inference_figure_MCMC}
\alias{generate_inference_figure_optim}
\title{Generate a figure of the estimates and credible intervals of a
  correlation matrix}
\usage{
generate_inference_figure_MCMC(stanfit, Rho.true, type = "marginal.hpd",
  conf = 0.95, ylimits = NULL, ln.Rho = TRUE, thin = 1,
  Rho.comp = NULL, segment_size = 2, text.size = 18,
  axis.ticks.x = TRUE, axis.text.size.x = NULL, verbose = FALSE,
  num_level = 0)

generate_inference_figure_optim(Fit, stanfit, fn, Rho.true, epsilon = 1e-06,
  ylimits = NULL, text.size = 20, verbose = FALSE, num_level = 0)
}
\arguments{
\item{stanfit}{An object of class \code{stanfit}.}

\item{Rho.true}{The true correlation matrix}

\item{type}{One of \code{"marginal.centered"} or \code{"marginal.hpd"},
indicating whether centeral or highest posterior density intervals are
desired}

\item{ylimits}{A vector of length two giving the lower and upper limits of
the y-axis}

\item{Rho.comp}{An optional correlation matrix in the composition}

\item{segment_size}{The size of the segments drawn in the plot}

\item{text.size}{The size of the plot text to use}

\item{axis.ticks.x}{A boolean indicating whether to put the coordinate
values on the x-axis.}

\item{axis.text.size.x}{If specified when \code{x.axis.ticks} is \code{TRUE},
then gives the text size of the x-axis tick labels.}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}

\item{Fit}{Either a \code{stanfit} or a list with element \code{par} being
a named vector of parameter estimates}

\item{fn}{The function whose gradient will be estimated.  It must take two
parameters: the first, a vector of parameter values; the second, a
\code{stanfit} object.}

\item{epsilon}{The step size used in estimating the derivative}
}
\description{
Only \code{generate_inference_figure_MCMC} can generate credible intervals
  of arbitrary confidence; \code{generate_inference_figure_optim} only
  produces 95\% credible intervals.
}

