% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_convergence_diagnostics.R
\name{get_convergence_diagnostics}
\alias{get_convergence_diagnostics}
\title{Get several convergence diagnostics for each parameter/chain from the fit.}
\usage{
get_convergence_diagnostics(samples, stan.output, calc_raftery = TRUE,
  calc_heidelberg = TRUE, geweke_fracs = c(0.1, 0.5), verbose = FALSE,
  num_level = 0)
}
\arguments{
\item{samples}{A set of samples extracted from a \code{stanfit} object with
\code{permuted=FALSE}}

\item{stan.output}{An list that contains the element "Fit", which is a
stanfit object}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}

\item{calculate_raftery}{Boolean: should the raftery diagnostic be calculated?}
}
\value{
A named list where each element is a matrix of parameters (rows) by
  convergence diagnostics (columns). The name of the element represents
  the type of convergence diagnostic
}
\description{
Get several convergence diagnostics for each parameter/chain from the fit.
}

