% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_convergence_diagnostics.R
\name{get_convergence_diagnostics_table}
\alias{get_convergence_diagnostics_table}
\title{Get the convergence diagnostics as a table}
\usage{
get_convergence_diagnostics_table(convergence_diagnostics,
  params_as_rows = TRUE, verbose = FALSE, num_level = 0)
}
\arguments{
\item{convergence_diagnostics}{A list of convergence diagnostics. Each element
should be a matrix of parameters (rows) by diagnostic statistic (columns)}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}

\item{output_file}{The name of the file to which to write all diagnostics}
}
\description{
Get the convergence diagnostics as a table
}

