% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_credible_intervals.R
\name{get_credible_intervals}
\alias{get_credible_intervals}
\title{Get credible intervals for the output according to different paradigms}
\usage{
get_credible_intervals(posterior_samples, list = FALSE,
  parameter.names = c("mu", "Sigma"), conf = 0.95, type = "marginal.hpd",
  verbose = FALSE, num_level = 0)
}
\arguments{
\item{posterior_samples}{A list of the posterior samples for each parameter.
Each element should be an array of the samples across all chains for a
parameter. The first dimension of the array should be the sample id; the
others, the parameter index.}

\item{list}{A boolean value indicating whether to return the quantiles as
an array or list.}

\item{parameter.names}{A list of parameters for which estimates or
quantiles will be returned}

\item{conf}{The width of the credible interval (conf * 100%).}

\item{type}{One of \code{"marginal.centered"} or \code{"marginal.hpd"},
indicating whether centeral or highest posterior density intervals are
desired}

\item{verbose}{Print informative statements as the function executes?xo}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\value{
Returns a list of the intervals for each parameter in
  \code{parameter.names}.  If \code{list=TRUE}, then each parameter is a
  list, with elements \code{lower} and \code{upper} being the lower and
  upper bounds, respectively. If \code{list=FALSE}, then each parameter is
  an array with the first dimension being the bounds and the remaining
  dimensions being the parameter index or indices.
}
\description{
Get credible intervals for the output according to different paradigms
}

