% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-variance-get_gradient.R
\name{get_gradient}
\alias{get_gradient}
\title{Estimate the gradient of a multivariate function at a particular domain
  value}
\usage{
get_gradient(fn, param.u, Fit, epsilon = 1e-06)
}
\arguments{
\item{fn}{The function whose gradient will be estimated.  It must take two
parameters: the first, a vector of parameter values; the second, a
\code{stanfit} object.}

\item{param.u}{The parameter values at which the gradient will be estimated}

\item{Fit}{A \code{stanfit} object (to be used as an argument to \code{fn}).}

\item{epsilon}{The step size used in estimating the derivative}
}
\description{
Estimate the gradient of a multivariate function at a particular domain
  value
}
\examples{
myfn <-
function(param.u, Fit){
    param.c <- constrain_pars(Fit, param.u)

    fn.param <- c(param.c$ln_mu,
                  param.c$ln_Rho[lower.tri(param.c$ln_Rho)],
                  sqrt(diag(param.c$ln_Sigma)))
    names(fn.param) <- c(rep("ln_mu", length(param.c$ln_mu)),
                         paste0("ln_Rho.", which(lower.tri(param.c$ln_Rho))),
                         rep("ln_sigma", length(diag(param.c$ln_Sigma))))
    return(fn.param)
}
}

