% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-figures-generate_inference_figure.R
\name{get_inference_df}
\alias{get_inference_df}
\title{Get the data frame for the inference plot from the correlation or covariance
  matrices}
\usage{
get_inference_df(Sigma.true, Sigma.median, Sigma.ci, Sigma.comp = NULL,
  corr = TRUE, verbose = FALSE, num_level = 0)
}
\arguments{
\item{Sigma.true}{The true value of the correlation or variance/covariance matrix}

\item{Sigma.median}{The estimated value of the correlation or variance/covariance matrix (assumed to be the posterior median)}

\item{Sigma.ci}{A list with two values: \code{lower} is the lower pointwise
limits of the 95\% credible interavls and \code{upper} is the upper
pointwise limits of the 95\% credible intervals.  Both elements are
matrices.}

\item{Sigma.comp}{An optional matrix of observed correlation values in the
composition}

\item{corr}{Boolean: Are these correlation matrices?}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Get the data frame for the inference plot from the correlation or covariance
  matrices
}

