% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-figures-generate_inference_figure.R
\name{get_inference_MCMC_df}
\alias{get_inference_MCMC_df}
\title{Get the data frames used for plotting MCMC inference figure}
\usage{
get_inference_MCMC_df(stanfit, Rho.true, type = "marginal.hpd", conf = 0.95,
  Rho.comp = NULL, ln.Rho = TRUE, thin = 1, verbose = FALSE,
  num_level = 0)
}
\arguments{
\item{stanfit}{An object of class \code{stanfit}.}

\item{Rho.true}{The true correlation matrix}

\item{type}{One of \code{"marginal.centered"} or \code{"marginal.hpd"},
indicating whether centeral or highest posterior density intervals are
desired}

\item{Rho.comp}{An optional correlation matrix in the composition}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Get the data frames used for plotting MCMC inference figure
}

