% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-run-get_IVs.R
\name{get_IVs}
\alias{get_IVs}
\title{Get initial values drawn from the prior}
\usage{
get_IVs(chains, data, verbose = FALSE, num_level = 0)
}
\arguments{
\item{chains}{The number of chains to run}

\item{data}{A data object, which must be a list with elements:
\itemize{
  \item \code{nu} The prior mean of mu
  \item \code{Lambda} The prior variance/covariance matrix of mu
  \item \code{eta} The parameter for the LKJ prior on the correlation
    matrix
  \item \code{alpha} The vector of shape parameters for the gamma prior on
    sigma
  \item \code{beta} The vector of scale parameters for the gamma prior on
    sigma
  \item \code{P} The number of features in the dataset
}}

\item{verbose}{Print informative statements as the function executes?xo}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Get initial values drawn from the prior
}

