% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IW_stats.R
\name{get_IW_stats}
\alias{get_IW_stats}
\title{Calculate summary statistics (mean, mode and variance) for an
  inverse-Wishart distribution}
\usage{
get_IW_stats(S, v)
}
\arguments{
\item{S}{Scale matrix \eqn{(p \times p)}{(p x p)}.}

\item{v}{Degrees of freedom (scalar).}
}
\value{
A named list: \code{Mean} is the mean, \code{Variance} is the
  variance, and \code{Mode} is the mode.
}
\description{
Calculate summary statistics (mean, mode and variance) for an
  inverse-Wishart distribution
}

