% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/melt_two_corr_mat.R
\name{get_melt_dataset}
\alias{get_melt_dataset}
\title{Generate a melted correlation dataset from a matrix using either a dataset
  or a correlation matrix}
\usage{
get_melt_dataset(mat, cor_mat = FALSE)
}
\arguments{
\item{mat}{Either a correlation matrix, or a dataset from which the
correlation will be calculated. If a dataset, samples must be rows.}

\item{cor_mat}{Boolean: is \code{mat} a correlation matrix?}
}
\value{
A melted data frame (as from \code{\link{reshape2::melt}})
}
\description{
Generate a melted correlation dataset from a matrix using either a dataset
  or a correlation matrix
}

