% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_min_width.R
\name{get_min_width}
\alias{get_min_width}
\title{Get credible intervals for the output according to different paradigms}
\usage{
get_min_width(posterior_samples, parameter.names = "ln_Rho", null_value = 0,
  type = "marginal.hpd", precision = 0.01, verbose = FALSE,
  num_level = 0)
}
\arguments{
\item{posterior_samples}{A list of the posterior samples for each parameter.
Each element should be an array of the samples across all chains for a
parameter. The first dimension of the array should be the sample id; the
others, the parameter index.}

\item{parameter.names}{A list of parameters for which estimates or
quantiles will be returned}

\item{null_value}{The null value to which to compare the credible interval.
Must be the same for all parameters.}

\item{type}{One of \code{"marginal.centered"} or \code{"marginal.hpd"},
indicating whether centeral or highest posterior density intervals are
desired}

\item{precision}{The precision used to estimate the minimum width.}

\item{verbose}{Print informative statements as the function executes?xo}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\value{
Returns a list of the minimum CI widths that include
  \code{null_value} for each parameter in \code{parameter.names}.
}
\description{
Get credible intervals for the output according to different paradigms
}

