% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-plot-get_plot_data.R
\name{get_plot_data}
\alias{get_Matrix_to_plot}
\alias{get_gamma_to_plot}
\alias{get_plot_data}
\alias{get_vec_to_plot}
\title{Get either a sample of vectors or matrices formatted for plotting}
\usage{
get_vec_to_plot(vec.sample, verbose = FALSE, num_level = 0)

get_Matrix_to_plot(Matrix.sample, diag = TRUE, verbose = FALSE,
  num_level = 0)

get_gamma_to_plot(alpha_vec, beta_vec, verbose = FALSE, num_level = 0)
}
\arguments{
\item{vec.sample}{A list of vectors}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}

\item{Matrix.sample}{A list of matrices}

\item{diag}{Boolean: include the diagonal elements of the matrix samples?}

\item{alpha_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}

\item{beta_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}
}
\value{
A data frame of indices and sample values (in the case of
  \code{get_gamma_to_plot}, pdf values for a range are returned)
}
\description{
Get either a sample of vectors or matrices formatted for plotting
}

