% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_posterior_quantities.R
\name{get_posterior_quantities}
\alias{get_posterior_estimates}
\alias{get_posterior_quantiles}
\alias{get_posterior_quantities}
\title{Get posterior estimates and quantiles}
\usage{
get_posterior_estimates(posterior_samples, estimate_method = "mean",
  parameter.names = c("mu", "Sigma"))

get_posterior_quantiles(posterior_samples, probs, list = FALSE,
  parameter.names = c("mu", "Sigma"))
}
\arguments{
\item{posterior_samples}{A list of the posterior samples for each parameter.
Each element should be an array of the samples across all chains for a
parameter. The first dimension of the array should be the sample id; the
others, the parameter index.}

\item{estimate_method}{Either \code{"mean"} or \code{"median"}, indicating
which estimate to use.}

\item{parameter.names}{A list of parameters for which estimates or
quantiles will be returned}

\item{list}{A boolean value indicating whether to return the quantiles as
an array or list.}
}
\value{
A named list where each element is the estimate or quantiles for a
  parameter in \code{parameter.names} The quantiles for a parameter are
  either an array with the first dimension being the quantile and the
  others the parameter indices; or a list where each element is the
  quantile of a parameter.
}
\description{
Get posterior estimates and quantiles
}

