% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prior-plot-get_prior_plot.R
\name{get_prior_plot}
\alias{get_Matrix_plot}
\alias{get_prior_plot}
\alias{get_vec_plot}
\title{Get density plots for either matrix or vector parameters}
\usage{
get_Matrix_plot(Matrix_to_plot, xlim = NULL, Matrix_text = NULL,
  verbose = FALSE)

get_vec_plot(vec_to_plot, vec_text = NULL, sampled = TRUE,
  mat_plot = FALSE, verbose = FALSE)
}
\arguments{
\item{Matrix_to_plot}{A data frame with columns \code{i} and \code{k}
indicating index and column \code{x} indicating a value of the parameter for
that index.}

\item{xlim}{Optional limits on the domain values of plot}

\item{Matrix_text}{A list of summary statistics for each parameter index. This
should be in the same format as that returned by
\code{\link{get_Matrix_text}}}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{vec_to_plot}{A data frame with columns \code{i} indicating index and
column \code{x} indicating a value of the parameter for that index. May
optionally contain column \code{y}, in which case \code{y} is a pdf
evaluated at \code{x} (if \code{sampled=FALSE})}

\item{vec_text}{A list of summary statistics for each
parameter index. This should be in the same format as that returned by
\code{\link{get_vec_text}}.}

\item{sampled}{Boolean: is this an estimated density based on sample, or an
actual calculated density? (if \code{sampled=FALSE}, then \code{vec_to_plot}
must contain columns \code{x} and \code{y})}

\item{mat_plot}{Boolean: Make a matrix of plots with \code{facet_grid} or
wrap the plots with \code{facet_wrap}?}
}
\value{
Returns a \code{ggplot} object which is the final plot of the samples.
}
\description{
Get density plots for either matrix or vector parameters
}

