% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prior-plot-get_prior_plot_sets.R
\name{get_prior_plot_sets}
\alias{get_IW_prior_plots}
\alias{get_LKJ_prior_plots}
\alias{get_prior_plot_sets}
\title{Generate plots of the prior distributions for all parameters for a few
  special prior cases}
\usage{
get_IW_prior_plots(n, v, S, nu, Lambda, gamma_mu = FALSE,
  lognorm_mu = FALSE, transform = "none", add.stats = FALSE, digits = 2,
  center.stat = "mean", verbose = FALSE)

get_LKJ_prior_plots(n, eta, alpha_vec, beta_vec, nu, Lambda, gamma_mu = FALSE,
  lognorm_mu = FALSE, transform = "none", add.stats = FALSE, digits = 2,
  center.stat = "mean", plot.Sigma = FALSE, verbose = FALSE)
}
\arguments{
\item{n}{The number of samples to use in generating the plots}

\item{v}{Degrees of freedom (scalar).}

\item{S}{Scale matrix \eqn{(p \times p)}{(p x p)}.}

\item{nu}{mean parameter for multivariate normal distribution}

\item{Lambda}{variance-covariance matrix for the multivariate normal
distribution}

\item{gamma_mu}{Boolean: is mu a gamma distribution with parameters
\code{nu} and \code{diag(Lambda)}?}

\item{lognorm_mu}{Boolean: exponentiate the mu samples after drawing?}

\item{transform}{one of \code{"none"}, \code{"normal:lognormal"}, or
\code{"lognormal:normal"}, indicating a transformation of the parameters
to be made. This allows priors corresponding to the parameter
transformation to be plotted.}

\item{add.stats}{Boolean: add text of summary statistics to the plots?}

\item{digits}{The number of digits to which to round summary statistics
(if plotted)}

\item{center.stat}{One of \code{"mean"} or \code{"median"}; the summary
statistic of central tendency to be plotted.}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{eta}{The scaling parameter of the LKJ distribution; must be > 1
(eta=1 means the distribution is uniform over d by d correlation matrices)}

\item{alpha_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}

\item{beta_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}

\item{plot.Sigma}{Boolean: Should the prior distribution of
\code{Sigma = diag(sigma) \%*\% R \%*\% diag(sigma)} be plotted?}
}
\value{
A named list: \code{.plot} are the \code{ggplot} objects, which can
  be printed; \code{.data} are the data objects.
}
\description{
Generate plots of the prior distributions for all parameters for a few
  special prior cases
}

