% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior_stats_to_print.R
\name{get_prior_stats_to_print}
\alias{get_prior_stats_to_print}
\title{Get summary statistics for each parameter that can be printed.}
\usage{
get_prior_stats_to_print(param.samples)
}
\arguments{
\item{param.samples}{A list of parameter samples, which can be scalar or
vector or matrix.}
}
\value{
A list of median and IQR for each parameter.
}
\description{
Get summary statistics for each parameter that can be printed.
}
\examples{
get_prior_stats_to_print(rnorm(10, 0, 1))

}

