% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-output-get_pseudo_p_values.R
\name{get_pseudo_p_values}
\alias{get_pseudo_p_values}
\title{A function to get pseudo p-values based on posterior credible intervals}
\usage{
get_pseudo_p_values(posterior_samples, parameter.names = "ln_Rho",
  null.value = 0, step.size = 0.01, verbose = FALSE)
}
\arguments{
\item{posterior_samples}{A list of the posterior samples for each parameter.
Each element should be an array of the samples across all chains for a
parameter. The first dimension of the array should be the sample id; the
others, the parameter index.}

\item{parameter.names}{A list of parameters for which estimates or
quantiles will be returned}

\item{null.value}{The null value to which the credible interval is being
compared}

\item{step.size}{The step size by which to increment the interval to
calculate the pseudo p-value.}
}
\value{
A list of pseudo p-values for each parameter in
  \code{parameter.names}
}
\description{
A function to get pseudo p-values based on posterior credible intervals
}

