% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_stan_calc.R
\name{get_R_from_stan}
\alias{get_R_from_stan}
\title{Get as an R object any vectors or matrices printed by stan}
\usage{
get_R_from_stan(stan_str)
}
\arguments{
\item{stan_str}{A string of a single line of printed output from
  \code{\link[rstan]{sampling}} which contains a vector or matrix.}
}
\value{
Returns the vector or matrix as an R vector or matrix
}
\description{
Get as an R object any vectors or matrices printed by stan
}
\examples{
stan_str1 <- "mu = [-0.689748,-0.836567,0.815985]"
stan_str2 <- "Rho = [[1,-0.0601858,-0.708844],[-0.0601858,1,-0.385467],[-0.708844,-0.385467,1]]"
stan_str3 <- "mu_log_prob = -3.67753"

get_R_from_stan(stan_str1)
get_R_from_stan(stan_str2)
get_R_from_stan(stan_str3)
}

