% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stan_calc.R
\name{get_stan_lp}
\alias{get_stan_lp}
\title{Get the numeric value of the log-posterior from a stan string}
\usage{
get_stan_lp(stan_str)
}
\arguments{
\item{stan_str}{A string of a single line of printed output from
\code{\link[rstan]{sampling}} which contains a log-posterior value}
}
\description{
Get the numeric value of the log-posterior from a stan string
}
\examples{
stan_str <- "Adding L; lp__ = -5.87068"
get_stan_lp(stan_str)

}

