% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prior-plot-get_text_to_plot.R
\name{get_text_to_plot}
\alias{get_Matrix_text}
\alias{get_gamma_text}
\alias{get_text_to_plot}
\alias{get_vec_text}
\title{get_text_to_plot}
\usage{
get_Matrix_text(Matrix.sample, digits = 2, diag = TRUE,
  center.stat = "mean", verbose = FALSE)

get_vec_text(vec.sample, digits = 2, center.stat = "mean",
  verbose = FALSE)

get_gamma_text(alpha_vec, beta_vec, digits = 2, verbose = FALSE)
}
\arguments{
\item{Matrix.sample}{A list of matrices}

\item{digits}{The number of digits to which to round summary statistics
(if plotted)}

\item{diag}{Boolean: include the diagonal elements of the matrix samples?}

\item{center.stat}{One of \code{"mean"} or \code{"median"}; the summary
statistic of central tendency to be plotted.}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{vec.sample}{A list of vectors}

\item{alpha_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}

\item{beta_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}
}
\value{
A list of summary statistics for each plot (determined by unique
  indicies of the parameter). Each element of the list is a data frame with
  columns \code{i} (and \code{k} in the case of \code{get_Matrix_text})
  indicating the index, \code{stat} indicating the value of the statistic, and
  \code{stat.type} indicating the name of the statistic (e.g., median or
  variance)
}
\description{
get_text_to_plot
}

