% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trans-variance-get_trans_variance.R
\name{get_trans_variance}
\alias{get_trans_variance}
\title{Get the variance for a transformation of the parameters}
\usage{
get_trans_variance(hessian.mat, fn, param.u, Fit, epsilon = 1e-05)
}
\arguments{
\item{hessian.mat}{The Hessian matrix for the original parameters}

\item{fn}{The function whose gradient will be estimated.  It must take two
parameters: the first, a vector of parameter values; the second, a
\code{stanfit} object.}

\item{param.u}{The parameter values at which the gradient will be estimated}

\item{Fit}{A \code{stanfit} object (to be used as an argument to \code{fn}).}

\item{epsilon}{The step size used in estimating the derivative}
}
\description{
Get the variance for a transformation of the parameters
}

