% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_corrmat_figure.R
\name{get_two_corrmat_df}
\alias{get_two_corrmat_df}
\title{Get a data frame for use with \code{make_two_corrmat_figure} from two
  matrices}
\usage{
get_two_corrmat_df(Rho.lower, Rho.upper, truestatus_mat = NA,
  verbose = FALSE, num_level = 0)
}
\arguments{
\item{Rho.lower}{The lower correlation matrix}

\item{Rho.upper}{The upper correlation matrix; must match \code{Rho.lower} in dimension}

\item{truestatus_mat}{An optional matrix indicating by Boolean values which
elements of Rho.lower and Rho.upper should be non-zero.  The dimensions
must match those of Rho.lower and Rho.upper.}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Get a data frame for use with \code{make_two_corrmat_figure} from two
  matrices
}

