% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_from_output.R
\name{grep_from_output}
\alias{grep_from_output}
\title{Get elements of a vector corresponding to a particular pattern}
\usage{
grep_from_output(print.output, pattern)
}
\arguments{
\item{print.output}{A vector of strings}

\item{pattern}{character string containing a \link{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr} and \code{gregexpr}.}
}
\value{
A vector of strings, each one of which contains \code{pattern}
}
\description{
Get elements of a vector corresponding to a particular pattern
}

