% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-output-get_min_width.R
\name{initialize_min_width}
\alias{initialize_min_width}
\title{Initializes a min_width list with names as parameter.names and
  value in each entry of vector or matrix, as appropriate (determined by
  samples_dim)}
\usage{
initialize_min_width(parameter.names, samples_dim, value, verbose = verbose,
  num_level = 0)
}
\description{
Initializes a min_width list with names as parameter.names and
  value in each entry of vector or matrix, as appropriate (determined by
  samples_dim)
}

