% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.symmetric}
\alias{is.symmetric}
\title{Check if a square matrix is symmetric}
\usage{
is.symmetric(mat)
}
\arguments{
\item{mat}{The square matrix to be checked}
}
\description{
Check if a square matrix is symmetric
}

