% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_stan_calc.R
\name{ll_components}
\alias{get_alpha}
\alias{get_mu.star}
\alias{get_sigmasq.star}
\alias{inc1}
\alias{inc2}
\alias{inc3}
\alias{inc4}
\alias{ll_components}
\alias{norm.constant}
\title{Calculate the components of the log-likelihood of C}
\usage{
norm.constant(i, C)

get_alpha(i, mu, C)

get_sigmasq.star(Sigma, C)

get_mu.star(i, mu, Sigma, C)

inc1(Sigma, C)

inc2(Sigma)

inc3(i, mu, Sigma, C)

inc4(i, mu, Sigma, C)
}
\arguments{
\item{i}{The index of the subject whose log-likelihood is calculated.}

\item{C}{The data matrix: rows indicate samples, columns indicate features.}

\item{mu}{The current value of the normal mean.}

\item{Sigma}{The current value of the normal variance-covariance matrix.}
}
\description{
Calculate the components of the log-likelihood of C
}

