% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-output-make_samples_list.R
\name{make_samples_list}
\alias{make_samples_list}
\title{Generate a list of samples when they have been extracted from a
  \code{stanfit} object with \code{permuted=FALSE}.}
\usage{
make_samples_list(samples, thin = 1, concatenate.chains = FALSE,
  verbose = FALSE)
}
\arguments{
\item{samples}{A set of samples extracted from a \code{stanfit} object with
\code{permuted=FALSE}}

\item{thin}{A positive integer specifying the period for saving samples; defaults to 1.}

\item{concatenate.chains}{Boolean: should all the chains be concatenated
together, or left separately?}

\item{verbose}{\code{TRUE} or \code{FALSE}: flag indicating whether 
     to print intermediate output from Stan on the console, which might
     be helpful for model debugging.}
}
\value{
A named list of each parameter from the \code{stanfit} object. The
  elements are either an array with the first dimension the iteration, the
  second the chain, and the third the parameter index; or an array with the
  first dimension the iteration (across all chains) and the remaining
  dimensions the parameter index (as in \code{\link[rstan]{extract}} when
  \code{permuted=TRUE}).
}
\description{
Generate a list of samples when they have been extracted from a
  \code{stanfit} object with \code{permuted=FALSE}.
}

