% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_corrmat_figure.R
\name{make_two_corrmat_figure}
\alias{make_two_corrmat_figure}
\title{Make a heatmap comparing two matrices when they are in data frame format.}
\usage{
make_two_corrmat_figure(Rho.df, title = NA, verbose = FALSE,
  categories = c("Truth", "Estimate"), corr_label = expression(rho["jj'"]),
  include_text = TRUE, text.size = 20, corr.text.size = 6,
  legend.position = "horizontal", num_level = 0)
}
\arguments{
\item{Rho.df}{The data frame; each row is an (i,j) element of one of the two
correlation matrices.  It must have the following columns:
\itemize{
  \item \code{Var1, y}: the (i,j) indices as factors
  \item \code{cor.value}: the value of the correlation of the (i,j)th
    element
}
It can also have the optional column \code{truestatus}, which is TRUE if
i and j are supposed to be associated and NA if not.}

\item{title}{The title of the plot}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{categories}{A character vector of length two. The first element
indicates the label for the lower triangle; the second indicates the label
for the upper triangle.}

\item{corr_label}{The label for the legend (can be an expression or a
string).}

\item{include_text}{Boolean: should the values of the correlation matrix be
added as text to the plot?}

\item{text.size}{The base size of the text for \code{ggplot} (in pt size).}

\item{legend.position}{One of either \code{"horizontal"} or
\code{"vertical"}; indicates whether the legends will be positioned at the
bottom (\code{"horizontal"}) or left (\code{"vertical"}) of the plot.}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Make a heatmap comparing two matrices when they are in data frame format.
}

