% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-figures-plot_convergence_diagnostics.R
\name{manipulate_convergence_diagnostics}
\alias{filter_convergence_diagnostics}
\alias{get_convergence_diagnostics_df}
\alias{manipulate_convergence_diagnostics}
\alias{tidy_convergence_diagnostics}
\title{Manipulate the convergence diagnostics to get useful tables}
\usage{
tidy_convergence_diagnostics(convergence_diagnostics_table, verbose = FALSE,
  num_level = 0)

filter_convergence_diagnostics(convergence_diagnostics_table_tidy,
  verbose = FALSE, num_level = 0)

get_convergence_diagnostics_df(convergence_diagnostics_table, filter = TRUE,
  verbose = FALSE, num_level = 0)
}
\arguments{
\item{convergence_diagnostics_table}{A table of convergence diagnostics;
columns are the diagnostics, rows are the parameters}

\item{verbose}{Print informative statements as the function executes?xo}

\item{num_level}{The number of the level (will determine the number of spaces to add)}

\item{convergence_diagnostics_table_tidy}{A tidy version of the
convergence diagnostics table as output by
\code{tidy_convergence_diagnostics}}

\item{filter}{Boolean: perform the filtering?}
}
\description{
Manipulate the convergence diagnostics to get useful tables
}

