% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/melt_two_corr_mat.R
\name{melt_two_corr_mat}
\alias{melt_two_corr_mat}
\title{Generate a melted dataset by combining two correlation matrices}
\usage{
melt_two_corr_mat(Rho.lower, Rho.upper, verbose = FALSE)
}
\arguments{
\item{Rho.lower}{The lower correlation matrix}

\item{Rho.upper}{The upper correlation matrix; must match \code{Rho.lower} in dimension}

\item{verbose}{Boolean: print informative statements as
function runs?}
}
\value{
A data frame with the following columns: \code{Var1} and \code{Var2}
  indicating the column/row indices of the correlation matrix;
  \code{cor.value} indicating the value of the correlation at those indices;
  \code{triangle} with values \code{"lower"} for the lower triangle elements
  and \code{"upper"} for the upper triangle elements.
}
\description{
Generate a melted dataset by combining two correlation matrices
}
\examples{
Rho.lower <- matrix(c(1, 0, 0, 0.5, 0,
                      0, 1, 0, 0, 0,
                      0, 0, 1, 0, 0,
                      0.5, 0, 0, 1, -0.5,
                      0, 0, 0, -0.5, 1), ncol=5)
X <- rmvnorm(5, rep(0, 5), Rho.lower)
Rho.upper <- cov2cor(X \%*\% t(X))
melt_two_corr_mat(Rho.lower, Rho.upper)
}

