% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{mycapture}
\alias{mycapture}
\title{A function that captures any printed output from an evaluated R expression
  and returns it as an element of a list}
\usage{
mycapture(..., file = NULL)
}
\arguments{
\item{...}{The expression to be evaluated}

\item{file}{An optional string or connection to which to write the output}
}
\value{
A list; \code{output} is the actual output of \code{...}, and
  \code{print.output} is the printed output from evaluating \code{...}.
}
\description{
A function that captures any printed output from an evaluated R expression
  and returns it as an element of a list
}

