% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_two_corr_mat.R
\name{mymerge}
\alias{mymerge}
\title{A function to merge two melted datasets}
\usage{
mymerge(x, y)
}
\arguments{
\item{x, y}{melted datasets, each with columns \code{Var1} and \code{Var2}}
}
\value{
A combined data frame combined by combinations of \code{Var1} and
  \code{Var2}
}
\description{
A function to merge two melted datasets
}
\examples{
x <- reshape2::melt(diag(5))
y <- reshape2::melt(diag(5) + 0.01 - diag(rep(5, 0.01)))
mymerge(x, y)
}

