% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newpage}
\alias{newpage}
\alias{spike_newpage}
\alias{spur_newpage}
\title{Add a LaTeX page break to a \code{knitr} report if the evaluation condition is met.}
\usage{
spike_newpage(eval.spike)

spur_newpage(eval.spur)
}
\arguments{
\item{eval.spike}{Whether to evaluate the spiked-in dataset}

\item{eval.spur}{Whether to evaluate the high spurious correlation dataset}
}
\description{
Add a LaTeX page break to a \code{knitr} report if the evaluation condition is met.
}

