% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_convergence_diagnostics.R
\name{output_convergence_diagnostics}
\alias{output_convergence_diagnostics}
\title{Write the convergence diagnostics to a file}
\usage{
output_convergence_diagnostics(convergence_diagnostics, output_file,
  params_as_rows = TRUE, verbose = FALSE, num_level = 0)
}
\arguments{
\item{convergence_diagnostics}{A list of convergence diagnostics. Each element
should be a matrix of parameters (rows) by diagnostic statistic (columns)}

\item{output_file}{The name of the file to which to write all diagnostics}

\item{params_as_rows}{Boolean; if TRUE, the rows of the output file will
be the parameters and the columns the diagnostic statistics, while if FALSE
the transpose will be output.}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Write the convergence diagnostics to a file
}

