% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_conditional_posterior.R
\name{plot_conditional_posterior}
\alias{plot_conditional_posterior}
\title{Generate a conditional posterior plot based on a Stan fit}
\usage{
plot_conditional_posterior(Fit, estimates, r.idx, r = NULL, ln = FALSE,
  verbose = FALSE)
}
\arguments{
\item{Fit}{A \code{stanfit} object}

\item{estimates}{The posterior estimates (usually posterior median)}

\item{r.idx}{A vector of length two indicating the indices of the element in
the correlation matrix whose plot will be generated}

\item{r}{The values of the correlation element over which to increment}

\item{ln}{Boolean: If TRUE, the parameter \code{ln_Rho} is plotted; if false,
the parameter \code{Rho} is plotted.}

\item{verbose}{Boolean: print informative statements as
function runs?}
}
\description{
Generate a conditional posterior plot based on a Stan fit
}

