% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-figures-plot_convergence_diagnostics.R
\name{plot_convergence_diagnostics}
\alias{plot_convergence_diagnostics}
\title{Tidy the convergecne diagnostics table}
\usage{
plot_convergence_diagnostics(convergence_diagnostics_table, filter = TRUE,
  important_stats = c("Rhat", "Geweke_z"), important_params = c("L",
  "sigma", "mu", "ln_Rho", "lp__"), verbose = FALSE, num_level = 0)
}
\arguments{
\item{convergence_diagnostics_table}{A table of convergence diagnostics;
columns are the diagnostics, rows are the parameters}

\item{filter}{Boolean: perform the filtering?}

\item{important_stats}{A vector of the names of statistics that will be
plotted; used to filter stat names}

\item{important_params}{A vector of the names of the key parameters to
look at; used to make an indicator of important variables.}

\item{verbose}{Print informative statements as the function executes?xo}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Tidy the convergecne diagnostics table
}

