% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prior-plot-plot_priors.R
\name{plot_priors}
\alias{plot_R_prior}
\alias{plot_Sigma_prior}
\alias{plot_mu_prior}
\alias{plot_priors}
\alias{plot_sigma_prior}
\title{Generate prior plots for several different possible types of priors}
\usage{
plot_Sigma_prior(Sigma.sample, add.stats = FALSE, digits = 2,
  center.stat = "mean", verbose = FALSE)

plot_mu_prior(mu.sample, add.stats = FALSE, digits = 2,
  center.stat = "mean", verbose = FALSE)

plot_sigma_prior(sigma.sample = NULL, alpha_vec = NULL, beta_vec = NULL,
  mat_plot = TRUE, add.stats = FALSE, digits = 2, center.stat = "mean",
  verbose = FALSE)

plot_R_prior(R.sample, add.stats = FALSE, digits = 2,
  center.stat = "mean", verbose = FALSE)
}
\arguments{
\item{Sigma.sample}{A list of variance-covariance matrices}

\item{add.stats}{Boolean: add text of summary statistics to the plots?}

\item{digits}{The number of digits to which to round summary statistics
(if plotted)}

\item{center.stat}{One of \code{"mean"} or \code{"median"}; the summary
statistic of central tendency to be plotted.}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{mu.sample}{A list of vectors}

\item{sigma.sample}{A list of vectors; either \code{sigma_sample} or BOTH
\code{alpha_vec} and \code{beta_vec} must be provided.}

\item{alpha_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}

\item{beta_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}

\item{mat_plot}{Boolean: Make a matrix of plots with \code{facet_grid} or
wrap the plots with \code{facet_wrap}?}

\item{R.sample}{A list of correlation matrices}
}
\value{
A named list: \code{plot} is a \code{ggplot} object; \code{data} is
  the data frame which was used to generate the plot
}
\description{
Generate prior plots for several different possible types of priors
}

