% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-figures-plot_stan_acf.R
\name{plot_stan_acf}
\alias{plot_stan_acf}
\alias{plot_stan_acf_set}
\title{Generate acf plots from a stanfit object}
\usage{
plot_stan_acf_set(Fit, NumChains, p, ln.params = FALSE, thin = 1,
  verbose = FALSE, num_level = 0)

plot_stan_acf(samples.list, NumChains, p, param.name, thin, is.matrix,
  plot.lower = TRUE, plot.diag = FALSE, verbose = FALSE, num_level = 0)
}
\arguments{
\item{Fit}{A \code{stanfit} object}

\item{NumChains}{The number of chains used in the fitting}

\item{p}{The dimension of the correlation matrix}

\item{ln.params}{Boolean: generate plots of lognormal parameters?}

\item{thin}{A positive integer specifying the period for saving samples; defaults to 1.}

\item{verbose}{\code{TRUE} or \code{FALSE}: flag indicating whether 
     to print intermediate output from Stan on the console, which might
     be helpful for model debugging.}

\item{num_level}{The number of the level (will determine the number of spaces to add)}

\item{samples.list}{An object as returned by \code{make_samples_list}.}

\item{is.matrix}{Boolean: is the parameter a matrix?}

\item{plot.lower}{Boolean: is the matrix symmetric?}

\item{plot.diag}{Boolean: include the diagnoal elements of a matrix?}
}
\value{
ACF plots are generated and printed on a per-parameter-element and
  per-chain basis.  Each parameter element is in a different plot; multiple
  chains are stacked side-by-side in a single plot.
}
\description{
Generate acf plots from a stanfit object
}

