% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes-model-print_model.R
\docType{data}
\name{printStanModel}
\alias{printStanModel}
\title{stanModel with added print statements}
\format{\preformatted{ chr "data \{\\n    int<lower=0> P; // number of elements of the composition\\n    int<lower=0> N; // the number of subjects sampled\\n  "| __truncated__
}}
\usage{
printStanModel
}
\description{
stanModel with added print statements
}
\keyword{datasets}

