% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-output-get_convergence_diagnostics.R
\name{read_in_convergence_diagnostics}
\alias{read_in_convergence_diagnostics}
\title{Read in convergence diagnostics from a file}
\usage{
read_in_convergence_diagnostics(input_file, delim = "\\t", verbose = FALSE,
  num_level = 0)
}
\arguments{
\item{input_file}{The name of the file with the convergence diagnostics.
These are assumed to have parameter names as the first column and
the convergence diagnostics on the remaining columns}

\item{delim}{Single character used to separate fields within a record.}
}
\description{
Read in convergence diagnostics from a file
}

