% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlkj.R
\name{rlkj}
\alias{rlkj}
\title{Generate a random sample from the LKJ distribution}
\usage{
rlkj(d, eta = 1, cholesky = FALSE, permute = !cholesky)
}
\arguments{
\item{d}{The dimension of the correlation matrix}

\item{eta}{The scaling parameter of the LKJ distribution; must be > 1
(eta=1 means the distribution is uniform over d by d correlation matrices)}

\item{cholesky}{Boolean: return the cholesky decomposition?}
}
\description{
This function is based on code from
\url{https://groups.google.com/forum/#!msg/stan-users/3gDvAs_qwN8/Xpgi2rPlx68J}.
}

