% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_banocc.R
\name{run_banocc}
\alias{run_banocc}
\title{Runs BAnOCC to fit the model and generate appropriate convergence metrics
  and inference.}
\usage{
run_banocc(bayes_model, C, nu, Lambda, alpha, beta, eta,
  cores = getOption("mc.cores", 1L), chains = 4, iter = 2000,
  warmup = floor(iter/2), thin = 1, init = "random", control = NULL,
  sd_mean = NULL, sd_var = NULL, conf_alpha = 0.05,
  get_min_width = FALSE, verbose = FALSE, num_level = 0)
}
\arguments{
\item{bayes_model}{The compiled stan model (as with
\code{stan_model(model_code = bayesStanModel)}).}

\item{C}{The dataset (rows=samples, columns=features), which is nxp}

\item{nu}{The prior mean for mu; vectors of length less than p will be
recycled.}

\item{Lambda}{The prior variance-covariance for mu (must be
positive-definite with dimension pxp where p=number of features in C), or
a vector of length p of variances for mu. If a vector of length less than p
is given, it will be recycled.}

\item{alpha, beta}{Vectors of parameters for the gamma prior on the standard
deviations. They must be of equal length. If a vector of length less than p
is given, it will be recycled.}

\item{eta}{scale parameter of LKJ distribution; must be >= 1.}

\item{cores}{Number of cores to use when executing the chains in parallel,
    which defaults to 1 but we recommend setting the \code{mc.cores} option 
    to be as many processors as the hardware and RAM allow (up to the 
    number of chains).}

\item{chains}{A positive integer specifying number of chains; defaults to 4.}

\item{iter}{A positive integer specifying how many iterations for each chain
    (including warmup). The default is 2000.}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin)
    iterations.  If step-size adaptation is on (which it is by default), this 
    also controls the number of iterations for which adaptation is run (and
    hence the samples should not be used for inference). The number of warmup
    should not be larger than \code{iter} and the default is \code{iter/2}.}

\item{thin}{A positive integer specifying the period for saving samples; defaults to 1.}

\item{init}{One of digit \code{0}, string \code{"0"} or \code{"random"}, 
    a function that returns a list, or a list of initial parameter values with
    which to indicate how the initial values of parameters are specified. 
    \code{"0"}: initialize all to be zero on the unconstrained support; 
    \code{"random"}: randomly generated;
    \code{list}: a list of lists equal in length to the number of chains 
     (parameter \code{chains}), where each list in the list of lists specifies the initial
      values of parameters by name for the corresponding chain.
    \code{function}: a function that returns a list for specifying 
      the initial values of parameters for a chain. The function
      can take an optional parameter \code{chain_id}. }

\item{control}{See the argument \code{control} of function
    \code{\link{stan}}.}

\item{sd_mean, sd_var}{Vectors of means and variances for the gamma priors on
the standard deviations. They must be of equal length. If a vector of length
less than p is given, it will be recycled.}

\item{get_min_width}{A boolean value: should the minimum CI width that
includes zero be calculated?}

\item{verbose}{\code{TRUE} or \code{FALSE}: flag indicating whether 
     to print intermediate output from Stan on the console, which might
     be helpful for model debugging.}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Runs BAnOCC to fit the model and generate appropriate convergence metrics
  and inference.
}

