% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-plot-sample_priors.R
\name{sample_prior}
\alias{sample_IW_prior}
\alias{sample_lkj_prior}
\alias{sample_mu_prior}
\alias{sample_prior}
\alias{sample_sigma_prior}
\title{Sampling from various prior distributions}
\usage{
sample_IW_prior(n, S, v, verbose = FALSE, num_level = 0)

sample_lkj_prior(n, d, eta, verbose = FALSE, num_level = 0)

sample_mu_prior(n, nu, Lambda, lognorm_mu = FALSE, verbose = FALSE,
  num_level = 0)

sample_sigma_prior(n, alpha_vec, beta_vec, verbose = FALSE, num_level = 0)
}
\arguments{
\item{n}{number of samples to draw from the distribution}

\item{S}{Scale matrix \eqn{(p \times p)}{(p x p)}.}

\item{v}{Degrees of freedom (scalar).}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}

\item{d}{The dimension of the correlation matrix}

\item{eta}{The scaling parameter of the LKJ distribution; must be > 1
(eta=1 means the distribution is uniform over d by d correlation matrices)}

\item{nu}{mean parameter for multivariate normal distribution}

\item{Lambda}{variance-covariance matrix for the multivariate normal
distribution}

\item{lognorm_mu}{Boolean: exponentiate the mu samples after drawing?}

\item{alpha_vec, beta_vec}{vectors of alpha and beta parameters for a gamma
distribution (see \code{\link{rgamma}}). The vectors must be paired
(ie., the first element of \code{alpha_vec} and the first element of
\code{beta_vec} means that the first sigma is distributed
gamma(\code{alpha_vec[1]}, \code{beta_vec[1]}))}
}
\description{
Sampling from various prior distributions
}

