% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-run-show_mcmc_time.R
\name{show_mcmc_time}
\alias{show_mcmc_time}
\title{Display the total time it took to run each chain of a \code{stan} run}
\usage{
show_mcmc_time(Fit.all)
}
\arguments{
\item{Fit.all}{The captured output (using \code{\link{mycapture}}) of the
generation of a \code{stanfit} object}
}
\value{
prints the lines of printed output corresponding to the time
}
\description{
Display the total time it took to run each chain of a \code{stan} run
}

