% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-run-show_optim.R
\name{show_optim}
\alias{show_optim}
\alias{show_optim_end}
\alias{show_optim_exit}
\alias{show_optim_start}
\alias{show_selected_optim_output}
\title{Display relevant parts of the printed outputs from running
  \code{stan::optimizing}}
\usage{
show_optim_end(Fit.all)

show_optim_exit(Fit.all)

show_optim_start(Fit.all)

show_selected_optim_output(Fit.all, exit = TRUE, startEnd = TRUE)
}
\arguments{
\item{Fit.all}{The captured output (using \code{\link{mycapture}}) of the
generation of a \code{stanfit} object}

\item{exit}{Boolean: print the exit status of the run?}

\item{startEnd}{Boolean: print the starting and ending portions of the run?}
}
\description{
Display relevant parts of the printed outputs from running
  \code{stan::optimizing}
}

