% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-output-get_min_width.R
\name{update_min_width}
\alias{update_min_width}
\title{Update the min_width list. Replaces every element to update with width.}
\usage{
update_min_width(min_width, which_to_update, width, verbose = FALSE,
  num_level = 0)
}
\description{
Update the min_width list. Replaces every element to update with width.
}

