% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-output-get_min_width.R
\name{update_missing_min_width}
\alias{update_missing_min_width}
\title{Updated the missing_min_width list; elements are set to FALSE if (1)
  they've just had a width added or (2) they have a missing width b/c
  the CI width is always zero (eg., correlation diagonal elts.)}
\usage{
update_missing_min_width(missing_min_width, which_to_update, width,
  verbose = FALSE, num_level = 0)
}
\description{
Updated the missing_min_width list; elements are set to FALSE if (1)
  they've just had a width added or (2) they have a missing width b/c
  the CI width is always zero (eg., correlation diagonal elts.)
}

