% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{vec.equal}
\alias{vec.equal}
\title{Check if all the elements of a vector are the same within a certain
  tolerance}
\usage{
vec.equal(vec, tol = 1e-06)
}
\arguments{
\item{vec}{A numeric vector}

\item{tol}{The tolerance}
}
\description{
Check if all the elements of a vector are the same within a certain
  tolerance
}

