% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-plot-warn_exponent.R
\name{warn_exponent}
\alias{warn_exponent}
\title{Warn if any values are too large/small for exponentiation}
\usage{
warn_exponent(sample, verbose = FALSE)
}
\arguments{
\item{sample}{a list of samples from a particular distribution}

\item{verbose}{Boolean: print the indices and values of the samples that are
too large or small}
}
\description{
Warn if any values are too large/small for exponentiation
}
\examples{
warn_exponent(1)
warn_exponent(710)
warn_exponent(list(-800, 0, 800, 12, -1e-16))
warn_exponent(list(-800, 0, 800, 12, -1e-16), verbose=TRUE)
}

