% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banocc_model.R
\docType{data}
\name{banocc_model}
\alias{banocc_model}
\title{The stan model used in the Bayesian fit}
\format{An object of class \code{character} of length 1.}
\usage{
banocc_model
}
\value{
The BAnOCC model as a string to be compiled with
    \code{rstan::stan_model}
}
\description{
This is the literal model used for fitting in Stan
}
\examples{
data(compositions_null)
\dontrun{
  compiled_banocc_model <- rstan::stan_model(model_code = banocc_model)
}

}
\keyword{datasets}

