% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_model.R
\docType{data}
\name{bayesStanModel}
\alias{bayesStanModel}
\title{The stan model used in the Bayesian fit}
\format{An object of class \code{character} of length 1.}
\usage{
bayesStanModel
}
\description{
This is the literal model used for fitting in Stan
}
\keyword{datasets}

