% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_v}
\alias{cat_v}
\title{Cat output only if verbose}
\usage{
cat_v(s_string, verbose, num_level = 0)
}
\arguments{
\item{verbose}{Print informative statements as the function executes?xo}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\description{
Cat output only if verbose
}

