% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compositions_neg_spike}
\alias{compositions_neg_spike}
\title{Simulated compositional data with a negative count correlation}
\format{A data frame with 1000 rows (compositional samples) and 9
  variables (the features)}
\usage{
compositions_neg_spike
}
\value{
A data frame with 1000 compositional samples from 9 features,
  generated by dividing each row of \code{counts_neg_spike} by its sum.
}
\description{
These are the normalized data corresponding to \code{counts_neg_spike}. The
  count data have one negative feature correlation, but the compositional
  correlation strucutre should be different.
}
\keyword{datasets}

