% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compositions_null}
\alias{compositions_null}
\title{Simulated compositional data with no feature correlations}
\format{A data frame with 1000 rows (compositional samples) and 9
  variables (the features)}
\usage{
compositions_null
}
\value{
A data frame with 1000 compositional samples from 9 features,
  generated by dividing each row of \code{counts_null} by its sum.
}
\description{
These are the normalized samples corresponding to \code{counts_null}. They
  should have a similar (but not identical) correlation structure.
}
\keyword{datasets}

