% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compositions_pos_spike}
\alias{compositions_pos_spike}
\title{Simulated compositional data with a positive count correlation}
\format{A data frame with 1000 rows (compositional samples) and 9
  variables (the features)}
\usage{
compositions_pos_spike
}
\value{
A data frame with 1000 compositional samples from 9 features,
  generated by dividing each row of \code{counts_pos_spike} by its sum.
}
\description{
These are the normalized data corresponding to \code{counts_pos_spike}. The
  count data have one positive feature correlation, but the compositional
  correlation structure should be different.
}
\keyword{datasets}

