% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-plot-convert_prior_samples.R
\name{convert_prior_samples}
\alias{convert_prior_samples}
\alias{convert_prior_to_lognormal}
\alias{get_norm.prior.sample}
\title{Convert mean and variance-covariance samples to lognormal parameters or
  normal parameters}
\usage{
convert_prior_to_lognormal(Sigma.sample, mu.sample, verbose = FALSE,
  num_level = 0)

get_norm.prior.sample(Sigma.sample, mu.sample, verbose = FALSE,
  num_level = 0)
}
\arguments{
\item{Sigma.sample}{A list of variance-covariance matrices}

\item{mu.sample}{A list of mean vectors}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{num_level}{The number of the level (will determine the number of spaces to add)}
}
\value{
A named list of converted parameters: \code{_mean} is the converted
  mean; \code{_Var} the converted variance; \code{_corr} the converted
  correlation matrix; and \code{_sd} the converted standard deviation
  vectors.
}
\description{
Convert mean and variance-covariance samples to lognormal parameters or
  normal parameters
}
\examples{
convert_prior_to_lognormal(diag(4), rep(0, 4))
convert_prior_to_normal(diag(4), rep(1, 4))
}

