% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan-figures-generate_inference_figure.R
\name{generate_inference_figure}
\alias{generate_inference_figure}
\alias{generate_inference_figure_MCMC}
\alias{generate_inference_figure_optim}
\title{Generate a figure of the estimates and credible intervals of a
  correlation matrix}
\usage{
generate_inference_figure_MCMC(Fit, Rho.true, type = "marginal.hpd",
  conf = 0.95, ylimits = NULL, ln.Rho = TRUE, thin = 1,
  Rho.comp = NULL, segment_size = 2, verbose = FALSE)

generate_inference_figure_optim(Fit, stanfit, fn, Rho.true, epsilon = 1e-06,
  ylimits = NULL, verbose = FALSE)
}
\arguments{
\item{Fit}{Either a \code{stanfit} or a list with element \code{par} being
a named vector of parameter estimates}

\item{Rho.true}{The true correlation matrix}

\item{type}{One of \code{"marginal.centered"} or \code{"marginal.hpd"},
indicating whether centeral or highest posterior density intervals are
desired}

\item{ylimits}{A vector of length two giving the lower and upper limits of
the y-axis}

\item{segment_size}{The size of the segments drawn in the plot}

\item{verbose}{Boolean: print informative statements as
function runs?}

\item{stanfit}{An object of class \code{stanfit}.}

\item{fn}{The function whose gradient will be estimated.  It must take two
parameters: the first, a vector of parameter values; the second, a
\code{stanfit} object.}

\item{epsilon}{The step size used in estimating the derivative}
}
\description{
Only \code{generate_inference_figure_MCMC} can generate credible intervals
  of arbitrary confidence; \code{generate_inference_figure_optim} only
  produces 95\% credible intervals.
}

