% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-model-print_model.R
\docType{data}
\name{printStanModel}
\alias{printStanModel}
\title{stanModel with added print statements}
\format{An object of class \code{character} of length 1.}
\usage{
printStanModel
}
\description{
stanModel with added print statements
}
\keyword{datasets}

