% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_banocc.R
\name{run_banocc}
\alias{run_banocc}
\title{Runs BAnOCC to fit the model and generate appropriate convergence metrics
  and inference.}
\usage{
run_banocc(bayes_model, C, nu = rep(0, ncol(C)), Lambda = 10 *
  diag(ncol(C)), alpha = rep(1, ncol(C)), beta = rep(0.5, ncol(C)),
  eta = 1, cores = getOption("mc.cores", 1L), chains = 4, iter = 2000,
  warmup = floor(iter/2), thin = 1, init = NULL, control = NULL,
  sd_mean = NULL, sd_var = NULL, conf_alpha = 0.05,
  get_min_width = FALSE, verbose = FALSE, num_level = 0)
}
\arguments{
\item{bayes_model}{The compiled stan model (as with
\code{stan_model(model_code = bayesStanModel)}).}

\item{C}{The dataset (rows=samples, columns=features), which is nxp}

\item{nu}{The prior mean for mu; vectors of length less than p will be
recycled.}

\item{Lambda}{The prior variance-covariance for mu (must be
positive-definite with dimension pxp where p=number of features in C), or
a vector of length p of variances for mu. If a vector of length less than p
is given, it will be recycled.}

\item{alpha, beta}{Vectors of parameters for the gamma prior on the standard
deviations. They must be of equal length. If a vector of length less than p
is given, it will be recycled.}

\item{eta}{scale parameter of LKJ distribution; must be >= 1.}

\item{cores}{Number of cores to use when executing the chains in parallel,
    which defaults to 1 but we recommend setting the \code{mc.cores} option 
    to be as many processors as the hardware and RAM allow (up to the 
    number of chains).}

\item{chains}{A positive integer specifying number of chains; defaults to 4.}

\item{iter}{A positive integer specifying how many iterations for each chain
    (including warmup). The default is 2000.}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin)
    iterations.  If step-size adaptation is on (which it is by default), this 
    also controls the number of iterations for which adaptation is run (and
    hence the samples should not be used for inference). The number of warmup
    should not be larger than \code{iter} and the default is \code{iter/2}.}

\item{thin}{A positive integer specifying the period for saving samples; defaults to 1.}

\item{init}{The initial values as a list (see
\code{\link[rstan]{sampling}}). Default value is NULL, which means that
initial values are sampled from the priors.}

\item{control}{See the argument \code{control} of function
    \code{\link{stan}}.}

\item{sd_mean, sd_var}{Vectors of means and variances for the gamma priors on
the standard deviations. They must be of equal length. If a vector of length
less than p is given, it will be recycled.}

\item{conf_alpha}{The percentage of the posterior density outside the
credible interval. That is, a \code{1-conf_alpha} * 100\% credible
interval will be returned.}

\item{get_min_width}{A boolean value: should the minimum CI width that
includes zero be calculated?}

\item{verbose}{Print informative statements as the function executes?}

\item{num_level}{The number of the level (will determine the number of
spaces to add to verbose output)}
}
\description{
Runs BAnOCC to fit the model and generate appropriate convergence metrics
  and inference.
}

