% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_banocc.R
\name{run_banocc}
\alias{run_banocc}
\title{Runs BAnOCC to fit the model and generate appropriate convergence metrics
  and inference.}
\usage{
run_banocc(banocc_model, C, n = rep(0, ncol(C)), L = 10 * diag(ncol(C)),
  a = rep(1, ncol(C)), b = rep(0.5, ncol(C)), eta = 1,
  cores = getOption("mc.cores", 1L), chains = 4, iter = 50,
  warmup = floor(iter/2), thin = 1, init = NULL, control = NULL,
  sd_mean = NULL, sd_var = NULL, conf_alpha = 0.05,
  get_min_width = FALSE, calc_snc = FALSE, verbose = FALSE,
  num_level = 0)
}
\arguments{
\item{banocc_model}{The compiled stan model (as with
\code{stan_model(model_code = banocc_model)}).}

\item{C}{The dataset as a data frame or matrix, with the samples as the
rows and the features as the columns.}

\item{n}{The prior mean for m; vectors of length less than p will be
recycled.}

\item{L}{The prior variance-covariance for m (must be
positive-definite with dimension pxp where p=number of features in C),
or a vector of length p of variances for m. If a vector of length less
than p is given, it will be recycled.}

\item{a, b}{Vectors of parameters for the gamma prior on the standard
deviations. They must be of equal length. If a vector of length less
than p is given, it will be recycled.}

\item{eta}{scale parameter of LKJ distribution; must be >= 1.}

\item{cores}{Number of cores to use when executing the chains in parallel,
    which defaults to 1 but we recommend setting the \code{mc.cores} option 
    to be as many processors as the hardware and RAM allow (up to the 
    number of chains).}

\item{chains}{A positive integer specifying the number of Markov chains. 
    The default is 4.}

\item{iter}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin)
    iterations per chain. If step-size adaptation is on (which it is by default), 
    this also controls the number of iterations for which adaptation is run (and
    hence these warmup samples should not be used for inference). The number of 
    warmup iterations should not be larger than \code{iter} and the default is 
    \code{iter/2}.}

\item{thin}{A positive integer specifying the period for saving samples. 
    The default is 1, which is usually the recommended value.}

\item{init}{The initial values as a list (see
\code{\link[rstan]{sampling}} in the \code{rstan} package). Default
value is NULL, which means that initial values are sampled from the
priors.}

\item{control}{A named \code{list} of parameters to control the sampler's
    behavior. See the details in the documentation for the \code{control} argument
    in \code{\link{stan}}.}

\item{sd_mean, sd_var}{Vectors of means and variances for the gamma priors
on the standard deviations. They must be of equal length. If a vector
of length less than p is given, it will be recycled.}

\item{conf_alpha}{The percentage of the posterior density outside the
credible interval. That is, a \code{1-conf_alpha} * 100\% credible
interval will be returned.}

\item{get_min_width}{A boolean value: should the minimum CI width that
includes zero be calculated?}

\item{calc_snc}{Boolean: should the scaled neighborhood criterion be
calculated?}

\item{verbose}{Print informative statements as the function executes?}

\item{num_level}{The number of indentations to add to the output when
\code{verbose = TRUE}.}
}
\value{
Returns a named list with the following elements:
\describe{
  \item{\emph{Data}}{The data formatted as a named list that includes the
    input data and the prior parameters}

  \item{\emph{Fit}}{The \code{stanfit} object returned by the call to
    \code{\link[rstan]{sampling}}}

  \item{\emph{CI}}{The \code{1-conf_alpha} * 100\% credible intervals}

  \item{\emph{Estimates.median}}{The correlation estimates, which are the
    marginal posterior medians}

  \item{\emph{Min.width}}{Only present if the \code{get_min_width}
    argument is \code{TRUE}. The minimum CI width that includes zero for
    each correlation.}

  \item{\emph{SNC}}{Only present if the \code{calc_snc} argument is
    \code{TRUE}. The scaled neighborhood criterion for each correlation.}
}
}
\description{
Runs BAnOCC to fit the model and generate appropriate convergence metrics
  and inference.
}
\examples{
\dontrun{
  banocc_model <- rstan::stan_model(model_code=banocc::banocc_model)
  b_output <- run_banocc(C=compositions_null, banocc_model=banocc_model)
}

}
\seealso{
\code{vignette("banocc-vignette")} for more examples.
}

